deleted a number of old functions.
Added one called writeOldNetworks() that returns a list with the output of all the write***NodeNetwork() functions. Marked the write***NodeNetwork() functions as not to be exported.

Re-documented functions. the help files for some old functions are now missing some parameter definitions because they inherited them from functions that were deleted.